♻️ [desktop]: Extract quick pick module and add i18n support
- Move quick pick logic from lib.rs and windows.rs to new quick_pick.rs module
- Add get_local_locale command to read locale from Rust settings for quick pick windows
- Initialize i18n for quick pick windows with locale fallback
- Add new translation keys for quick pick UI labels and hints
- Update QuickPick and QuickPickResult Vue components to use translations
- Remove unused imports from lib.rs and windows.rs
