+++ apps/desktop/src/locales/en.json
@@ -103,6 +103,13 @@
   "main.quick_pick.no_commands": "No commands found",
   "main.quick_pick.configure": "Configure slash prompts",
   "main.quick_pick.result_title": "Quick Pick Result",
+  "main.quick_pick.prompt_label": "Prompt",
+  "main.quick_pick.selection_label": "Selection",
+  "main.quick_pick.selection_empty": "Selection is empty",
+  "main.quick_pick.hint_typed_prompt": "Press Enter to run your typed prompt",
+  "main.quick_pick.hint_select_command": "Select a command or press Enter to run",
+  "main.quick_pick.cancel_hint": "Press Esc to cancel",
+  "main.quick_pick.command_label": "Command",
   "main.status.processing": "Processing...",
   "main.action.close": "Close",
   "main.action.copy": "Copy",
