+++ apps/desktop/src/locales/jp.json
@@ -103,6 +103,13 @@
   "main.quick_pick.no_commands": "コマンドが見つかりません",
   "main.quick_pick.configure": "スラッシュコマンドを設定",
   "main.quick_pick.result_title": "クイックピック結果",
+  "main.quick_pick.prompt_label": "プロンプト",
+  "main.quick_pick.selection_label": "選択範囲",
+  "main.quick_pick.selection_empty": "選択範囲が空です",
+  "main.quick_pick.hint_typed_prompt": "Enter で入力したプロンプトを実行",
+  "main.quick_pick.hint_select_command": "コマンドを選択するか Enter で実行",
+  "main.quick_pick.cancel_hint": "Esc でキャンセル",
+  "main.quick_pick.command_label": "コマンド",
   "main.status.processing": "処理中...",
   "main.action.close": "閉じる",
   "main.action.copy": "コピー",
