+++ apps/desktop/src/locales/zh.json
@@ -103,6 +103,13 @@
   "main.quick_pick.no_commands": "未找到命令",
   "main.quick_pick.configure": "配置斜杠命令",
   "main.quick_pick.result_title": "Quick Pick 结果",
+  "main.quick_pick.prompt_label": "提示词",
+  "main.quick_pick.selection_label": "选区",
+  "main.quick_pick.selection_empty": "选区为空",
+  "main.quick_pick.hint_typed_prompt": "按 Enter 运行你输入的提示词",
+  "main.quick_pick.hint_select_command": "选择命令或按 Enter 运行",
+  "main.quick_pick.cancel_hint": "按 Esc 取消",
+  "main.quick_pick.command_label": "命令",
   "main.status.processing": "处理中...",
   "main.action.close": "关闭",
   "main.action.copy": "复制",
